Bump to 0.2.3
add 2 new flags:
* problem_nonstrict to yaml_parser_t
* indentless_map to yaml_emitter_t

indentless_map is for backcompat with yaml readers.

It changes the default indent behavior of map elements,
advance the indent the children of the map
according to the spec: 6.1 Indentation Spaces:

      "Each node must be indented further than its parent node.  All sibling nodes
      must use the exact same indentation level.  However the content of each
      sibling node may be further indented independently."
      http://yaml.org/spec/1.2/spec.html#id2777534

This make it readable by YAML.pm (perl),
and behaves the same as all other YAML writers.

problem_nonstrict is to skip aborting on some reader errors:
"control characters are not allowed"
The error is set, but parsing is continued. This allows optionally
accepting invalid string values, copy-as-is, avoiding data-loss.